+++ apm-agent-plugins/apm-servlet-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -14,7 +14,7 @@ co.elastic.apm.agent.servlet.JakartaAsyncInstrumentation$JakartaStartAsyncInstru
 co.elastic.apm.agent.servlet.JakartaAsyncInstrumentation$JakartaAsyncContextInstrumentation
 co.elastic.apm.agent.servlet.JavaxRequestStreamRecordingInstrumentation
 co.elastic.apm.agent.servlet.JakartaRequestStreamRecordingInstrumentation
-co.elastic.apm.agent.servlet.service_name.Servlet2PlusInstrumentation$JavaxInitServiceNameInstrumentation
-co.elastic.apm.agent.servlet.service_name.Servlet2PlusInstrumentation$JakartaInitServiceNameInstrumentation
-co.elastic.apm.agent.servlet.service_name.Servlet3PlusInstrumentation$JavaxInitServiceNameInstrumentation
-co.elastic.apm.agent.servlet.service_name.Servlet3PlusInstrumentation$JakartaInitServiceNameInstrumentation
+co.elastic.apm.agent.servlet.servicename.InitServiceNameInstrumentation$JavaxInitServiceNameInstrumentation
+co.elastic.apm.agent.servlet.servicename.InitServiceNameInstrumentation$JakartaInitServiceNameInstrumentation
+co.elastic.apm.agent.servlet.servicename.ServletContainerInitializerServiceNameInstrumentation$JavaxInitServiceNameInstrumentation
+co.elastic.apm.agent.servlet.servicename.ServletContainerInitializerServiceNameInstrumentation$JakartaInitServiceNameInstrumentation
